Reorganize unit tests onto a test_<layer>_<unit>.py naming convention
Phase 2 of the unit-test refactor: file motion only.  Function bodies are
carried over verbatim, so this commit should read as pure movement.

Renames:
  test_structs.py           -> test_plugin_abcs.py  (it tests the plugin ABCs)
  test_latin_hypercube.py   -> test_searches.py
  test_numpy_database.py    -> test_db_numpy.py
  test_obj_lib.py           -> test_lib_objectives.py
  test_const_lib.py         -> test_lib_constraints.py
  test_sims_simple.py       -> test_lib_simulations.py
  test_libe.py              -> test_extras_libe.py
  test_viz_without_dash.py  -> test_viz.py
  test_moop_setup.py        -> test_core_setup.py

Splits, one file per unit under test:
  test_moop_base.py    -> test_core_embed.py, test_core_surrogates.py,
                          test_core_evaluate.py
  test_moop_solve.py   -> test_core_iterate.py, test_core_solve.py
  test_moop_grads.py   -> test_core_grads.py, and test_MOOP_solve_with_grads
                          joins test_core_solve.py
  test_moop_db_ops.py  -> test_core_db_ops.py, test_core_checkpoint.py

Every file's __main__ block is now pytest.main([__file__]) rather than a
hand-maintained list of calls, so direct `python3 test_x.py` execution keeps
working without having to be kept in sync as tests are added or split.

All 30 files now also pass standalone, which was not previously true: the
jax_enable_x64 config was set at module scope in test_moop_grads.py and
leaked into every other file at collection time, so single-file runs used
32-bit precision while full-suite runs used 64-bit.  conftest.py now owns it.

Coverage gate: 4085/4085 baseline lines still executed.  97 tests unchanged.
