+++ parmoo/tests/unit_tests/test_core_iterate.py
@@ -1,3 +1,9 @@
+""" Unit tests for the MOOP iterate/filterBatch/updateAll solve loop steps.
+"""
+
+
+import pytest
+
 
 def test_MOOP_iterate():
     """ Test the MOOP class's iterator in objectives.py.
@@ -317,187 +323,5 @@ def test_MOOP_iterate():
         assert (xi["x2"] == "0")
 
 
-def test_MOOP_solve():
-    """ Test the MOOP class's solver in objectives.py.
-
-    Perform a test of the MOOP solver class by minimizing a 5 variable,
-    biobjective convex function s.t. $x in [0, 1]^n$.
-
-    The correctness of the solutions is difficult to assert , but we can
-    assert  that the efficient points map onto the Pareto front, as
-    expected.
-
-    """
-
-    from parmoo import MOOP
-    from parmoo.searches import LatinHypercube
-    from parmoo.surrogates import GaussRBF
-    from parmoo.acquisitions import UniformWeights, RandomConstraint
-    from parmoo.optimizers import LocalSurrogate_PS, GlobalSurrogate_BFGS
-    import numpy as np
-    import pytest
-
-    # Initialize two simulation groups with 1 output each
-    s1 = {'m': 1,
-          'hyperparams': {},
-          'search': LatinHypercube,
-          'sim_func': lambda x: [np.linalg.norm([x[f"x{key}"]
-                                                 for key in [1, 2, 3, 4]])],
-          'surrogate': GaussRBF,
-          'search_budget': 20}
-    s2 = {'m': 1,
-          'hyperparams': {},
-          'search': LatinHypercube,
-          'sim_func': lambda x: [np.linalg.norm([x[f"x{key}"] - 1
-                                                 for key in [1, 2, 3, 4]])],
-          'surrogate': GaussRBF,
-          'search_budget': 20}
-    # Create a MOOP with 4 design variables and 2 simulations
-    moop1 = MOOP(LocalSurrogate_PS, hyperparams={'opt_budget': 100,
-                                                 'np_random_gen': 0})
-    for i in range(4):
-        moop1.addDesign({'lb': 0.0, 'ub': 1.0})
-    moop1.addSimulation(s1, s2)
-    # Now add 2 objectives
-    def f1(x, sim): return sim["sim1"]
-    def f2(x, sim): return sim["sim2"]
-    moop1.addObjective({'obj_func': f1},
-                       {'obj_func': f2})
-    # Add 3 acquisition functions
-    for i in range(3):
-        moop1.addAcquisition({'acquisition': UniformWeights})
-    # Try to solve several invalid problems/budgets to test error handling
-    with pytest.raises(ValueError):
-        moop1.solve(-1)
-    with pytest.raises(TypeError):
-        moop1.solve(2.0)
-    # Solve the MOOP with 6 iterations
-    moop1.solve(6)
-    soln = moop1.getPF()
-    # Assert that solutions were found
-    assert (soln.size > 0)
-    # Assert that the x_vals and f_vals match
-    for i in range(soln.shape[0]):
-        assert (np.linalg.norm(np.array([s1['sim_func'](soln[i]),
-                                         s2['sim_func'](soln[i])]).flatten() -
-                               np.array([soln['f1'][i], soln['f2'][i]]))
-                < 1.e-8)
-    # Create new single objective toy problem
-    s3 = {'m': 1,
-          'sim_func': lambda x: [x["x1"] + x["x2"]],
-          'surrogate': GaussRBF,
-          'search': LatinHypercube,
-          'hyperparams': {'search_budget': 10}}
-    s4 = {'m': 1,
-          'sim_func': lambda x: [x["x3"] + x["x4"]],
-          'surrogate': GaussRBF,
-          'search': LatinHypercube,
-          'hyperparams': {'search_budget': 20}}
-    moop2 = MOOP(LocalSurrogate_PS, hyperparams={'np_random_gen': 0})
-    for i in range(4):
-        moop2.addDesign({'lb': 0.0, 'ub': 1.0})
-    moop2.addSimulation(s3, s4)
-    # Now add 1 objective
-    def f3(x, sim): return sim["sim1"][0] + sim["sim2"][0]
-    moop2.addObjective({'obj_func': f3})
-    # Add 3 acquisition functions
-    for i in range(3):
-        moop2.addAcquisition({'acquisition': RandomConstraint})
-    # Solve the MOOP and extract the final database with 6 iterations
-    moop2.solve(6)
-    soln = moop2.getPF()
-    # Assert that solutions were found
-    assert (soln.size > 0)
-    # Assert that the x_vals and f_vals match
-    for i in range(soln.shape[0]):
-        assert (np.linalg.norm(np.array(s3['sim_func'](soln[i])) +
-                               np.array(s4['sim_func'](soln[i])) -
-                               soln['f1'][i]) < 0.00000001)
-
-    # Create a 3 objective toy problem, with no simulations
-    moop3 = MOOP(GlobalSurrogate_BFGS, hyperparams={'np_random_gen': 0})
-    for i in range(4):
-        moop3.addDesign({'lb': 0.0, 'ub': 1.0})
-
-    # Now add the three objectives
-    def f4(x, sim):
-        return np.linalg.norm([x["x1"] - 1, x["x2"], x["x3"], x["x4"]]) ** 2.0
-
-    def g4(x, sim):
-        dx = {"x1": 2 * x["x1"] - 2,
-              "x2": 2 * x["x2"],
-              "x3": 2 * x["x3"],
-              "x4": 2 * x["x4"]}
-        return dx, {}
-
-    def f5(x, sim):
-        return np.linalg.norm([x["x1"], x["x2"] - 1, x["x3"], x["x4"]]) ** 2.0
-
-    def g5(x, sim):
-        dx = {"x1": 2 * x["x1"],
-              "x2": 2 * x["x2"] - 2,
-              "x3": 2 * x["x3"],
-              "x4": 2 * x["x4"]}
-        return dx, {}
-
-    def f6(x, sim):
-        return np.linalg.norm([x["x1"], x["x2"], x["x3"] - 1, x["x4"]]) ** 2.0
-
-    def g6(x, sim):
-        dx = {"x1": 2 * x["x1"],
-              "x2": 2 * x["x2"],
-              "x3": 2 * x["x3"] - 2,
-              "x4": 2 * x["x4"]}
-        return dx, {}
-
-    moop3.addObjective({'obj_func': f4, 'obj_grad': g4},
-                       {'obj_func': f5, 'obj_grad': g5},
-                       {'obj_func': f6, 'obj_grad': g6})
-    # Add 3 acquisition functions
-    for i in range(3):
-        moop3.addAcquisition({'acquisition': UniformWeights})
-    # Solve the MOOP and extract the final database with 6 iterations
-    moop3.solve(6)
-    soln = moop3.getPF()
-    # Assert that solutions were found
-    assert (soln.size > 0)
-    # Assert that the x_vals and f_vals match
-    sim = np.zeros(0)
-    for i in range(soln.shape[0]):
-        assert (np.linalg.norm(np.array([f4(soln[i], sim),
-                                         f5(soln[i], sim),
-                                         f6(soln[i], sim)]
-                                        ).flatten() -
-                               np.array([soln[i]["f1"], soln[i]["f2"],
-                                         soln[i]["f3"]])) < 0.00000001)
-
-    # Create a 3 objective toy problem, with no simulations and 1 categorical
-    moop4 = MOOP(GlobalSurrogate_BFGS, hyperparams={'np_random_gen': 0})
-    for i in range(3):
-        moop4.addDesign({'lb': 0.0, 'ub': 1.0})
-    moop4.addDesign({'des_type': "categorical", 'levels': 3})
-    moop4.addObjective({'obj_func': f4, 'obj_grad': g4},
-                       {'obj_func': f5, 'obj_grad': g5},
-                       {'obj_func': f6, 'obj_grad': g6})
-    # Add 3 acquisition functions
-    for i in range(3):
-        moop4.addAcquisition({'acquisition': UniformWeights})
-    # Solve the MOOP and extract the final database with 6 iterations
-    moop4.solve(6)
-    soln = moop4.getPF()
-    # Assert that solutions were found
-    assert (soln.size > 0)
-    # Assert that the x_vals and f_vals match
-    sim = np.zeros(0)
-    for i in range(soln.shape[0]):
-        assert (np.linalg.norm(np.array([f4(soln[i], sim),
-                                         f5(soln[i], sim),
-                                         f6(soln[i], sim)]
-                                        ).flatten() -
-                               np.array([soln[i]["f1"], soln[i]["f2"],
-                                         soln[i]["f3"]])) < 0.00000001)
-
-
 if __name__ == "__main__":
-    test_MOOP_iterate()
-    test_MOOP_solve()
+    raise SystemExit(pytest.main([__file__]))
